ckb-auth-types, remove all dependencies
