+++ ckb-auth-types/Cargo.toml
@@ -6,6 +6,3 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-log = { version = "0.4.17", default-features = false }
-ckb-std = "0.14.3"
-hex = { version = "0.4.3", default-features = false, features = ["alloc"]}
